Sobol4R-package

Tools to design experiments, compute Sobol sensitivity indices, and summarise stochastic responses inspired by the strategy described by Lebrun et al. (2021) doi:10.1016/j.ress.2021.107385 . Includes helpers to optimise toy models implemented in C++, visualise indices with uncertainty quantification, and derive reliability-oriented sensitivity measures based on failure probabilities. It is further detailed in Logosha, Maumy and Bertrand (2022) doi:10.1063/5.0246026 and (2023) doi:10.1063/5.0246024 or in Bertrand, Logosha and Maumy (2024) https://hal.science/hal-05371803, https://hal.science/hal-05371795 and https://hal.science/hal-05371798.

References

Elizaveta Logosha, Myriam Maumy, Frederic Bertrand; Confidence interval determination using discrete event simulations for real estate sales case. AIP Conf. Proc. 31 March 2025; 3182 (1): 100008. doi.org:10.1063/5.0246026.

Elizaveta Logosha, Myriam Maumy, Frédéric Bertrand; Sensitivity analysis of stochastic simulator in the case of sales date prediction. AIP Conf. Proc. 31 March 2025; 3182 (1): 100001. doi.org:10.1063/5.0246024

Frédéric Bertrand, Elizaveta Logosha, Myriam Maumy-Bertrand. Extension of sensitivity analysis to uncertainties in distribution parameters. 32nd Conference on Intelligent Systems for Molecular Biology, International Society for Computational Biology, Jul 2024, Montreal (QC), Canada. https://hal.science/hal-05371795

Frédéric Bertrand, Elizaveta Logosha, Myriam Maumy-Bertrand. Sobol4RV: Global Sensitivity Analysis in Several Random Settings. BioC 2024, BioConductor, Jul 2024, Grand Rapids, MI, United States. https://hal.science/hal-05371803

Frédéric Bertrand, Elizaveta Logosha, Myriam Maumy-Bertrand. Global Sensitivity Analysis in Several Random Settings. 2024 Joint Statistical Meetings, American Statistical Association, Aug 2024, Portland (OR), United States. https://hal.science/hal-05371798

See also

Author

Maintainer: Frederic Bertrand frederic.bertrand@lecnam.net (ORCID)

Authors:

Elizaveta Logosha elizaveta.logosha@utt.fr

Myriam Maumy-Bertrand myriam.maumy@ehesp.fr (ORCID)

Examples

ex1_results <- sobol_example_g_deterministic(n=100, nboot=10) 
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
#> Warning: extreme order statistics used as endpoints
print(ex1_results)
#> 
#> Call:
#> sensitivity::sobol(model = NULL, X1 = X1, X2 = X2, order = order,     nboot = nboot)
#> 
#> Model runs: 3700 
#> 
#> Sobol indices
#>           original         bias std. error  min. c.i. max. c.i.
#> X1     0.585090675  0.018126903 0.09405824  0.4178692 0.7204938
#> X2     0.211628471  0.009633229 0.18662941 -0.2527058 0.3590147
#> X3     0.075643436  0.034787113 0.16209177 -0.2347347 0.2680770
#> X4     0.023250862  0.025304945 0.13519051 -0.2196896 0.2149281
#> X5     0.002067305  0.021295934 0.14094288 -0.2447231 0.1781164
#> X6     0.002387349  0.021015813 0.14027207 -0.2398426 0.1774792
#> X7     0.003709273  0.018024314 0.14147879 -0.2434797 0.1853099
#> X8     0.002619155  0.020598742 0.13991390 -0.2423908 0.1771970
#> X1*X2  0.093057957 -0.024800619 0.17509720 -0.1101237 0.4662503
#> X1*X3  0.014274687 -0.032931339 0.14550054 -0.1677026 0.3018598
#> X1*X4  0.007123201 -0.021533158 0.16023513 -0.1882466 0.3114686
#> X1*X5 -0.004655849 -0.021228972 0.14109774 -0.1809365 0.2442021
#> X1*X6 -0.002508963 -0.021144863 0.14096659 -0.1784576 0.2457039
#> X1*X7 -0.003864528 -0.022276491 0.14116194 -0.1795775 0.2456577
#> X1*X8 -0.001841283 -0.020837445 0.14156475 -0.1784560 0.2469993
#> X2*X3  0.001911336 -0.023124162 0.13753322 -0.1580454 0.2452932
#> X2*X4  0.006056803 -0.021434949 0.14545495 -0.1815817 0.2499668
#> X2*X5 -0.003438179 -0.020703019 0.14109671 -0.1809424 0.2441975
#> X2*X6 -0.003051052 -0.020606153 0.14128191 -0.1798270 0.2440936
#> X2*X7 -0.001373232 -0.021064719 0.14158951 -0.1784634 0.2477255
#> X2*X8 -0.002866535 -0.021117924 0.14094765 -0.1782685 0.2454887
#> X3*X4 -0.003368247 -0.020221639 0.13980411 -0.1779483 0.2404287
#> X3*X5 -0.003020588 -0.020921569 0.14113075 -0.1794425 0.2451085
#> X3*X6 -0.002594781 -0.021016973 0.14133175 -0.1790731 0.2468932
#> X3*X7 -0.002909885 -0.021082523 0.14136739 -0.1798306 0.2461687
#> X3*X8 -0.003012063 -0.021060978 0.14130212 -0.1795975 0.2457053
#> X4*X5 -0.002547326 -0.020929039 0.14120563 -0.1790049 0.2456916
#> X4*X6 -0.002732773 -0.020946788 0.14115537 -0.1793299 0.2453462
#> X4*X7 -0.002602843 -0.021029066 0.14120961 -0.1793559 0.2455148
#> X4*X8 -0.002785757 -0.021146610 0.14115783 -0.1791784 0.2452370
#> X5*X6 -0.002727353 -0.021008306 0.14116461 -0.1791644 0.2453976
#> X5*X7 -0.002732033 -0.021002584 0.14115415 -0.1791388 0.2453386
#> X5*X8 -0.002723232 -0.021019009 0.14115462 -0.1791391 0.2453816
#> X6*X7 -0.002721261 -0.021005195 0.14116151 -0.1791487 0.2453913
#> X6*X8 -0.002716813 -0.021003607 0.14115667 -0.1791285 0.2453719
#> X7*X8 -0.002739390 -0.021014024 0.14115020 -0.1791300 0.2453452
ggplot(ex1_results)
#> Error in ggplot(ex1_results): could not find function "ggplot"
rm(ex1_results)