Fast Ishigami Test Function

C++ implementation of the Ishigami function that is widely used as a benchmark for Sobol sensitivity indices. The implementation is vectorised and therefore convenient for Monte Carlo experiments.

ishigami_model(x, a = 7, b = 0.1)

Arguments

x

Numeric matrix with three columns representing the inputs.

a

Numeric scalar controlling the nonlinear term.

b

Numeric scalar controlling the interaction term.

Value

Numeric vector of simulator outputs.

Examples

x <- matrix(runif(30, -pi, pi), ncol = 3)
ishigami_model(x)
#>  [1]  4.27101919  7.74296966 -1.62315376  0.02410397  5.87917536  1.07708949
#>  [7]  2.84310098  7.74317767  5.42819479  1.71344673