Summarise Sobol Indices

Compute compact summaries of the Sobol indices and their Monte Carlo variability. The function is intended to feed diagnostic plots.

summarise_sobol(result, probs = c(0.1, 0.5, 0.9), bootstrap = 200L)

Arguments

result

A sobol_result object.

probs

Numeric vector of probabilities used to report quantiles of the empirical bootstrap distribution.

bootstrap

Integer, number of bootstrap resamples used to quantify the estimator uncertainty.

Value

A data frame (class sobol_summary) with the requested statistics. Quantile columns are added when probs is not empty.

Examples

design <- sobol_design(n = 64, d = 3)
model <- function(x) ishigami_model(x)
sob <- sobol_indices(model, design, keep_samples = TRUE)
summarise_sobol(sob, probs = c(0.1, 0.9))
#>   parameter   first_order  total_order variance noise_variance    first_q01
#> 1        X1  0.0625096145 2.803088e-02 2.736324              0 -0.005282309
#> 2        X2  1.1378947693 1.016588e+00 2.736324              0  0.720635404
#> 3        X3 -0.0009182027 3.444694e-05 2.736324              0 -0.002813747
#>      total_q01   first_q09    total_q09
#> 1 2.243541e-02 0.121708978 3.539335e-02
#> 2 8.552551e-01 1.571546537 1.192745e+00
#> 3 2.120965e-05 0.000938778 4.787327e-05